feat: apps are fully installed
During EC2 instance initialization of the
Online Mixer instance we create a startup
script `install-apps.sh` which will run when
a user first logs in to Ubuntu Desktop. This
is ensured by the `install-apps.desktop`file
which will be deleted after the app
installations, so that it will run only once.
We also set real-time process permissions
for the user `ubuntu`.
And we ensure the Ubuntu instance is using
a German keyboard layout.