feat: the `config.json` includes now the channels
The `channels` config parameter is an array of string with the names of
the channels that are used during the jam session. These names will be
used to create the Jamulus ini files for the different Jamulus clients
as well as to name the channels of the Ardour mixing console and to
connect the correct Jamulus instance output to the dedicated Ardour
channel.