feat: full Ardour session with 10 channels
The session also includes a reverb auxiliary bus and every channel has
the following plugins added:
- An EQ
- A compressor
- An auxiliary send to the reverb bus

The reverb bus has a reverb plugin and an equalizer.
The master bus has a compressor plugin.
All channels and the reverb bus send their output to the master bus.
Only the Jamulus instance `MixAcoustic` is the input for all channels.