fix: dependencies
You can just run a Jamulus band server, or a band server with an audio
workstation. However, if you also want to include a Zoom server, you
must have a mixing server in place as well and it only makes sense to
have a Jamulus mixing server when you also have an audio workstation.